refactor(storage): inline cloud storage configuration
