+++ app/components/settings.tsx
@@ -10,24 +10,14 @@ import ClearIcon from "../icons/clear.svg";
 import LoadingIcon from "../icons/three-dots.svg";
 import EditIcon from "../icons/edit.svg";
 import EyeIcon from "../icons/eye.svg";
-import DownloadIcon from "../icons/download.svg";
-import UploadIcon from "../icons/upload.svg";
-import ConfigIcon from "../icons/config.svg";
-import ConfirmIcon from "../icons/confirm.svg";
-
-import ConnectionIcon from "../icons/connection.svg";
-import CloudSuccessIcon from "../icons/cloud-success.svg";
-import CloudFailIcon from "../icons/cloud-fail.svg";
 import {
   Input,
   List,
   ListItem,
   Modal,
-  PasswordInput,
   Popover,
   Select,
   showConfirm,
-  showToast,
 } from "./ui-lib";
 
 import { IconButton } from "./button";
@@ -47,17 +37,14 @@ import Locale, {
 } from "../locales";
 import { copyToClipboard, clientUpdate, semverCompare } from "../utils";
 import Link from "next/link";
-import { Path, RELEASE_URL, STORAGE_KEY, UPDATE_URL } from "../constant";
+import { Path, RELEASE_URL, UPDATE_URL } from "../constant";
 import { SearchService, usePromptStore } from "../store/prompt";
 import { ErrorBoundary } from "./error";
 import { InputRange } from "./input-range";
 import { useNavigate } from "react-router-dom";
 import { Avatar, AvatarPicker, WalletAccount } from "./emoji";
 import { getClientConfig } from "../config/client";
-import { useSyncStore } from "../store/sync";
 import { nanoid } from "nanoid";
-import { useSkillStore } from "../store/skill";
-import { ProviderType } from "../utils/cloud";
 import { TTSConfigList } from "./tts-config";
 import { RealtimeConfigList } from "./realtime-chat/realtime-config";
 import {
@@ -296,402 +283,6 @@ function AccountItems() {
         />
       </ListItem>
     </List>
-  );
-}
-
-function CheckButton() {
-  const syncStore = useSyncStore();
-  const webdavEnvBaseUrl =
-    getClientConfig()?.webdavBackendBaseUrl?.trim() || "";
-  const webdavEnvPrefix = getClientConfig()?.webdavBackendPrefix?.trim() || "";
-  const webdavBaseUrl = syncStore.webdav.baseUrl || webdavEnvBaseUrl;
-  const webdavPrefix = syncStore.webdav.baseUrl.trim()
-    ? syncStore.webdav.prefix
-    : syncStore.webdav.prefix || webdavEnvPrefix;
-
-  const couldCheck = useMemo(() => {
-    return syncStore.cloudSync();
-  }, [syncStore]);
-
-  const [checkState, setCheckState] = useState<
-    "none" | "checking" | "success" | "failed"
-  >("none");
-
-  async function check() {
-    setCheckState("checking");
-    const valid = await syncStore.check();
-    setCheckState(valid ? "success" : "failed");
-  }
-
-  if (!couldCheck) return null;
-
-  return (
-    <IconButton
-      text={Locale.Settings.Sync.Config.Modal.Check}
-      bordered
-      onClick={check}
-      icon={
-        checkState === "none" ? (
-          <ConnectionIcon />
-        ) : checkState === "checking" ? (
-          <LoadingIcon />
-        ) : checkState === "success" ? (
-          <CloudSuccessIcon />
-        ) : checkState === "failed" ? (
-          <CloudFailIcon />
-        ) : (
-          <ConnectionIcon />
-        )
-      }
-    ></IconButton>
-  );
-}
-
-export function SyncConfigModal(props: { onClose?: () => void }) {
-  const syncStore = useSyncStore();
-  const webdavEnvBaseUrl =
-    getClientConfig()?.webdavBackendBaseUrl?.trim() || "";
-  const webdavEnvPrefix = getClientConfig()?.webdavBackendPrefix?.trim() || "";
-  const webdavBaseUrl = syncStore.webdav.baseUrl || webdavEnvBaseUrl;
-  const webdavPrefix = syncStore.webdav.baseUrl.trim()
-    ? syncStore.webdav.prefix
-    : syncStore.webdav.prefix || webdavEnvPrefix;
-
-  return (
-    <div className="modal-mask">
-      <Modal
-        title={Locale.Settings.Sync.Config.Modal.Title}
-        onClose={() => props.onClose?.()}
-        actions={[
-          <CheckButton key="check" />,
-          <IconButton
-            key="confirm"
-            onClick={props.onClose}
-            icon={<ConfirmIcon />}
-            bordered
-            text={Locale.UI.Confirm}
-          />,
-        ]}
-      >
-        <List>
-          <ListItem
-            title={Locale.Settings.Sync.Config.SyncType.Title}
-            subTitle={Locale.Settings.Sync.Config.SyncType.SubTitle}
-          >
-            <select
-              value={syncStore.provider}
-              onChange={(e) => {
-                syncStore.update(
-                  (config) =>
-                    (config.provider = e.target.value as ProviderType),
-                );
-              }}
-            >
-              {Object.entries(ProviderType).map(([k, v]) => (
-                <option value={v} key={k}>
-                  {k}
-                </option>
-              ))}
-            </select>
-          </ListItem>
-
-          <ListItem title={Locale.Settings.Sync.Config.AutoSync.Title}>
-            <input
-              type="checkbox"
-              checked={syncStore.autoSync}
-              onChange={(e) => {
-                syncStore.update(
-                  (config) => (config.autoSync = e.currentTarget.checked),
-                );
-              }}
-            ></input>
-          </ListItem>
-
-          {syncStore.autoSync ? (
-            <ListItem
-              title={Locale.Settings.Sync.Config.AutoSyncInterval.Title}
-            >
-              <input
-                type="number"
-                min={1}
-                max={120}
-                value={Math.max(
-                  1,
-                  Math.round(syncStore.autoSyncIntervalMs / 60000),
-                )}
-                onChange={(e) => {
-                  const minutes = Number(e.currentTarget.value);
-                  if (!Number.isFinite(minutes)) return;
-                  syncStore.update(
-                    (config) =>
-                      (config.autoSyncIntervalMs =
-                        Math.max(1, minutes) * 60000),
-                  );
-                }}
-              ></input>
-            </ListItem>
-          ) : null}
-
-          <ListItem
-            title={Locale.Settings.Sync.Config.Proxy.Title}
-            subTitle={Locale.Settings.Sync.Config.Proxy.SubTitle}
-          >
-            <input
-              type="checkbox"
-              checked={syncStore.useProxy}
-              onChange={(e) => {
-                syncStore.update(
-                  (config) => (config.useProxy = e.currentTarget.checked),
-                );
-              }}
-            ></input>
-          </ListItem>
-          {syncStore.useProxy ? (
-            <ListItem
-              title={Locale.Settings.Sync.Config.ProxyUrl.Title}
-              subTitle={Locale.Settings.Sync.Config.ProxyUrl.SubTitle}
-            >
-              <input
-                type="text"
-                value={syncStore.proxyUrl}
-                onChange={(e) => {
-                  syncStore.update(
-                    (config) => (config.proxyUrl = e.currentTarget.value),
-                  );
-                }}
-              ></input>
-            </ListItem>
-          ) : null}
-        </List>
-
-        <List>
-          <ListItem title={Locale.Settings.Sync.Config.WebDav.AuthType}>
-            <select
-              value={syncStore.webdav.authType}
-              onChange={(e) => {
-                syncStore.update(
-                  (config) =>
-                    (config.webdav.authType = e.target
-                      .value as typeof config.webdav.authType),
-                );
-              }}
-            >
-              <option value="basic">Basic</option>
-              <option value="ucan">UCAN</option>
-            </select>
-          </ListItem>
-        </List>
-
-        {syncStore.provider === ProviderType.WebDAV && (
-          <>
-            <List>
-              <ListItem
-                title={
-                  syncStore.webdav.authType === "ucan"
-                    ? Locale.Settings.Sync.Config.WebDav.UcanBaseUrl
-                    : Locale.Settings.Sync.Config.WebDav.BaseUrl
-                }
-                subTitle={Locale.Settings.Sync.Config.WebDav.BaseUrlSubTitle}
-              >
-                <input
-                  type="text"
-                  value={webdavBaseUrl}
-                  onChange={(e) => {
-                    syncStore.update(
-                      (config) =>
-                        (config.webdav.baseUrl = e.currentTarget.value),
-                    );
-                  }}
-                ></input>
-              </ListItem>
-              <ListItem
-                title={
-                  syncStore.webdav.authType === "ucan"
-                    ? Locale.Settings.Sync.Config.WebDav.UcanPrefix
-                    : Locale.Settings.Sync.Config.WebDav.Prefix
-                }
-                subTitle={Locale.Settings.Sync.Config.WebDav.PrefixSubTitle}
-              >
-                <input
-                  type="text"
-                  placeholder="/dav"
-                  value={webdavPrefix}
-                  onChange={(e) => {
-                    syncStore.update(
-                      (config) =>
-                        (config.webdav.prefix = e.currentTarget.value),
-                    );
-                  }}
-                ></input>
-              </ListItem>
-              {syncStore.webdav.authType === "basic" ? (
-                <>
-                  <ListItem title={Locale.Settings.Sync.Config.WebDav.UserName}>
-                    <input
-                      type="text"
-                      value={syncStore.webdav.username}
-                      onChange={(e) => {
-                        syncStore.update(
-                          (config) =>
-                            (config.webdav.username = e.currentTarget.value),
-                        );
-                      }}
-                    ></input>
-                  </ListItem>
-                  <ListItem title={Locale.Settings.Sync.Config.WebDav.Password}>
-                    <PasswordInput
-                      value={syncStore.webdav.password}
-                      onChange={(e) => {
-                        syncStore.update(
-                          (config) =>
-                            (config.webdav.password = e.currentTarget.value),
-                        );
-                      }}
-                    ></PasswordInput>
-                  </ListItem>
-                </>
-              ) : null}
-            </List>
-          </>
-        )}
-
-        {syncStore.provider === ProviderType.UpStash && (
-          <List>
-            <ListItem title={Locale.Settings.Sync.Config.UpStash.Endpoint}>
-              <input
-                type="text"
-                value={syncStore.upstash.endpoint}
-                onChange={(e) => {
-                  syncStore.update(
-                    (config) =>
-                      (config.upstash.endpoint = e.currentTarget.value),
-                  );
-                }}
-              ></input>
-            </ListItem>
-
-            <ListItem title={Locale.Settings.Sync.Config.UpStash.UserName}>
-              <input
-                type="text"
-                value={syncStore.upstash.username}
-                placeholder={STORAGE_KEY}
-                onChange={(e) => {
-                  syncStore.update(
-                    (config) =>
-                      (config.upstash.username = e.currentTarget.value),
-                  );
-                }}
-              ></input>
-            </ListItem>
-            <ListItem title={Locale.Settings.Sync.Config.UpStash.Password}>
-              <PasswordInput
-                value={syncStore.upstash.apiKey}
-                onChange={(e) => {
-                  syncStore.update(
-                    (config) => (config.upstash.apiKey = e.currentTarget.value),
-                  );
-                }}
-              ></PasswordInput>
-            </ListItem>
-          </List>
-        )}
-      </Modal>
-    </div>
-  );
-}
-
-function SyncItems() {
-  const syncStore = useSyncStore();
-  const chatStore = useChatStore();
-  const promptStore = usePromptStore();
-  const skillStore = useSkillStore();
-  const couldSync = useMemo(() => {
-    return syncStore.cloudSync();
-  }, [syncStore]);
-
-  const [showSyncConfigModal, setShowSyncConfigModal] = useState(false);
-
-  const stateOverview = useMemo(() => {
-    const sessions = chatStore.sessions;
-    const messageCount = sessions.reduce((p, c) => p + c.messages.length, 0);
-
-    return {
-      chat: sessions.length,
-      message: messageCount,
-      prompt: Object.keys(promptStore.prompts).length,
-      skill: Object.keys(skillStore.skills).length,
-    };
-  }, [chatStore.sessions, skillStore.skills, promptStore.prompts]);
-
-  return (
-    <>
-      <List>
-        <ListItem
-          title={Locale.Settings.Sync.CloudState}
-          subTitle={
-            syncStore.lastProvider
-              ? `${new Date(syncStore.lastSyncTime).toLocaleString()} [${
-                  syncStore.lastProvider
-                }]`
-              : Locale.Settings.Sync.NotSyncYet
-          }
-        >
-          <div style={{ display: "flex" }}>
-            <IconButton
-              aria={Locale.Settings.Sync.CloudState + Locale.UI.Config}
-              icon={<ConfigIcon />}
-              text={Locale.UI.Config}
-              onClick={() => {
-                setShowSyncConfigModal(true);
-              }}
-            />
-            {couldSync && (
-              <IconButton
-                icon={<ResetIcon />}
-                text={Locale.UI.Sync}
-                onClick={async () => {
-                  try {
-                    await syncStore.sync({ interactive: true });
-                    showToast(Locale.Settings.Sync.Success);
-                  } catch (e) {
-                    showToast(Locale.Settings.Sync.Fail);
-                    console.error("[Sync]", e);
-                  }
-                }}
-              />
-            )}
-          </div>
-        </ListItem>
-
-        <ListItem
-          title={Locale.Settings.Sync.LocalState}
-          subTitle={Locale.Settings.Sync.Overview(stateOverview)}
-        >
-          <div style={{ display: "flex" }}>
-            <IconButton
-              aria={Locale.Settings.Sync.LocalState + Locale.UI.Export}
-              icon={<UploadIcon />}
-              text={Locale.UI.Export}
-              onClick={() => {
-                syncStore.export();
-              }}
-            />
-            <IconButton
-              aria={Locale.Settings.Sync.LocalState + Locale.UI.Import}
-              icon={<DownloadIcon />}
-              text={Locale.UI.Import}
-              onClick={() => {
-                syncStore.import();
-              }}
-            />
-          </div>
-        </ListItem>
-      </List>
-
-      {showSyncConfigModal && (
-        <SyncConfigModal onClose={() => setShowSyncConfigModal(false)} />
-      )}
-    </>
   );
 }
 
@@ -995,8 +586,6 @@ export function Settings() {
               ></input>
             </ListItem>
           </List>
-
-          <SyncItems />
 
           <List>
             <ListItem
